+++ apps-api/backoffice/src/modules/profile/index.ts
@@ -14,6 +14,7 @@ import {
   GetProfileByIdParams,
   GetProfileByIdResponse,
   GetUserParticipationResponse,
+  GetUserRecommendationResponse,
   UpdateProfileBody,
   UpdateProfileResponse,
 } from './models'
@@ -26,6 +27,32 @@ export const ProfileController = new Elysia({
   tags: ['Profile'],
 })
   .use([AuthGuardPlugin, ProfileServicePlugin])
+  .get(
+    '/recommend',
+    async ({ user, profileService, status }) => {
+      const result = await profileService.getUserRecommendation(user.id)
+
+      if (result.isErr()) {
+        return mapErrorCodeToResponse(result.error, status)
+      }
+
+      return status(200, result.value)
+    },
+    {
+      requiredLocalUser: true,
+      response: {
+        200: GetUserRecommendationResponse,
+        ...createErrorSchema(
+          InternalErrorCode.UNAUTHORIZED,
+          InternalErrorCode.INTERNAL_SERVER_ERROR
+        ),
+      },
+      detail: {
+        summary: 'Get User Recommendations',
+        description: "Fetch the authenticated user's recommended users",
+      },
+    }
+  )
   .get(
     '/participation',
     async ({ user, profileService, status }) => {
@@ -64,7 +91,7 @@ export const ProfileController = new Elysia({
       return status(200, {
         id: result.value.id,
         name: result.value.name,
-        role: result.value.role,
+        roles: result.value.roles,
         profileImage: result.value.profileImage ?? undefined,
         numberOfFollowing: result.value.numberOfFollowing,
         point: result.value.point,