+++ apps-client/mobile/app/(tabs)/(feed)/_layout.tsx
@@ -7,9 +7,12 @@ export default function FeedLayout() {
       screenOptions={{ headerShown: false, animation: 'slide_from_right' }}
     >
       <Stack.Screen name="index" />
-      <Stack.Screen name="[feedId]" />
       <Stack.Screen name="hashtag/[hashtagId]" />
       <Stack.Screen name="announcement/index" />
+      <Stack.Screen name="user-suggestion" />
+      <Stack.Screen name="topic/[topicId]" />
+      <Stack.Screen name="topic/suggestion" />
+      <Stack.Screen name="[feedId]" />
     </Stack>
   )
 }