+++ apps-client/mobile/app/(tabs)/activity.tsx
@@ -4,8 +4,8 @@ import { H1 } from '@pple-today/ui/typography'
 
 export default function ActivityPage() {
   return (
-    <View className="flex-1 bg-base-bg-light flex flex-col justify-center items-center">
-      <H1 className="text-2xl font-anakotmai-medium py-4 text-base-text-medium">Coming Soon</H1>
+    <View className="pt-safe flex-1 bg-base-bg-light flex flex-col justify-center items-center">
+      <H1 className="text-2xl font-heading-semibold py-4 text-base-text-medium">Coming Soon</H1>
     </View>
   )
 }