+++ apps-client/mobile/components/linear-gradient.tsx
@@ -1,0 +1,5 @@
+import { LinearGradient } from 'expo-linear-gradient'
+import { cssInterop } from 'nativewind'
+
+cssInterop(LinearGradient, { className: 'style' })
+export { LinearGradient }