+++ packages/database/__generated__/prisma/sql/get_candidate_user.d.ts
@@ -1,0 +1,13 @@
+import * as $runtime from "../runtime/client"
+
+/**
+ * @param text
+ */
+export const get_candidate_user: (text: string) => $runtime.TypedSql<get_candidate_user.Parameters, get_candidate_user.Result>
+
+export namespace get_candidate_user {
+  export type Parameters = [text: string]
+  export type Result = {
+    user_id: string | null
+  }
+}