+++ packages/api-client/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @pple-today/api-client
 
+## 0.3.2
+
+### Patch Changes
+
+- [#95](https://github.com/PPLEThai/pple-today/pull/95) [`8ec7aeb`](https://github.com/PPLEThai/pple-today/commit/8ec7aeb678d9db9b56c15d28ec7330b59bba53a3) Thanks [@miello](https://github.com/miello)! - [[PPLE-357] [API] Lint script problem and profile image caching](https://linear.app/snts/issue/PPLE-357/api-lint-script-problem-and-profile-image-caching)
+
+- [#113](https://github.com/PPLEThai/pple-today/pull/113) [`a277f86`](https://github.com/PPLEThai/pple-today/commit/a277f86ffefc0d8da80b26b116337864f452b32c) Thanks [@miello](https://github.com/miello)! - [[PPLE-367] [API] Improve DX in fetch client interceptor](https://linear.app/snts/issue/PPLE-367/api-improve-dx-in-fetch-client-interceptor)
+
+- [#140](https://github.com/PPLEThai/pple-today/pull/140) [`63d3100`](https://github.com/PPLEThai/pple-today/commit/63d31001c929e6f0a471077411593b5527fa1163) Thanks [@Anon-136](https://github.com/Anon-136)! - [[PPLE-229] Follow Suggestion Integrate](https://linear.app/snts/issue/PPLE-229/follow-suggestion-integrate)
+  [[PPLE-231] Follow Suggestion List Integrate](https://linear.app/snts/issue/PPLE-231/follow-suggestion-list-integrate)
+  [[PPLE-236] Topic Suggestion List Page Integrate](https://linear.app/snts/issue/PPLE-236/topic-suggestion-list-page-integrate)
+  [[PPLE-233] Topic Suggestion Integrate](https://linear.app/snts/issue/PPLE-233/topic-suggestion-integrate)
+
+- [#96](https://github.com/PPLEThai/pple-today/pull/96) [`ca5323e`](https://github.com/PPLEThai/pple-today/commit/ca5323e773cf34cebb357972cd25a479f1c960a3) Thanks [@miello](https://github.com/miello)! - [[PPLE-358] [API] Allows type-only import from API](https://linear.app/snts/issue/PPLE-358/api-allows-type-only-import-from-api)
+
 ## 0.3.1
 
 ### Patch Changes